feat(core): virtualize release activity list
